refactor(core,cli): split cascade and tag modules below 500 LOC gate (WG-185)
Both cascade/mod.rs and tag/core.rs sat at exactly 500 LOC (the quality
gate limit). Extract CascadeConfig/TierResult/CascadeResult into
cascade/types.rs and search_by_tags into tag/core/search_ops.rs. Also fix a
pre-existing unused-variable warning in cascade/weights.rs tests.

Amp-Thread-ID: https://ampcode.com/threads/T-019f1cc7-c9a7-76e1-9e8e-75f5cf91a537
